Generic targeted refresh support
Updated the ems_refresher_mixin to support a pattern for targeted refresh.  This
was largely taken from the VMWare targeted refresh process.

The major difference with the targeted refresh approach is that refreshers
should collect inventory from the provider separate from parsing that inventory.
In many more recent refreshers, the inventory is collected as it's being parsed.
The problem this poses with targeted refresh is that it doesn't allow the
refresher to filter the data down to the specific targets that should be
refreshed.

This implementation allows "legacy" refreshers (those that don't yet support
targeted refresh) to continue to collect inventory while parsing it.  However,
as providers implement targeted refresh, they should adopt the process of
separately collecting inventory.

The other major change with this implementation is that now the VMWare refresher
uses the same ems_refersher_mixin that all other refreshers use.  This helps to
solidify the ability for the mixin to handle both non-targeted and targeted
refreshes.

The major caveat to this approach is that the logging from the old VMWare
refresher is slightly different from the ems_refresher_mixin.  If log analysis
tools exist that rely on the old logging, they will need to be updated to
examine the new logging labels (see updated Benchmark calls in this commit for
details).